requirements.txt
# ===========================================
# PYTHON PROJECT REQUIREMENTS - UNIFIED
# Generated: 2025-01-10
# ===========================================

# Core LangChain ecosystem
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.78
langchain-google-genai==1.0.0
langchain-openai==0.3.34
langchain-text-splitters==0.3.11

# LangGraph for workflow orchestration
langgraph==0.2.74

# Monitoring and observability
langsmith==0.4.32

# Vector database and embeddings
faiss-cpu==1.12.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
numpy==2.3.3

# AI/ML libraries
openai==2.1.0
tiktoken==0.11.0
pypdf==6.1.1

# Web and HTTP
aiohttp==3.12.15
httpx==0.28.1
httpx-sse==0.4.1
requests==2.32.5
requests-toolbelt==1.0.0

# Data processing and utilities
pandas>=2.0.0
scikit-learn>=1.3.0
psutil>=5.9.0

# Configuration and validation
pydantic==2.11.9
pydantic-settings==2.11.0
python-dotenv==1.1.1

# Logging and monitoring
structlog>=23.0.0

# Database and storage
SQLAlchemy==2.0.43

# Development and testing
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# Code quality and formatting
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Type checking and inspection
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2

# Async and concurrent programming
asyncio
aiofiles>=0.23.0

# Image and media processing (for multimodal RAG)
Pillow>=10.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Deployment and containerization
docker>=6.0.0

# ===========================================
# DEPENDENCY GROUPS (Optional - use with pip-tools or similar)
# ===========================================

# Production dependencies (above)
# Development dependencies (above)
# Testing dependencies (above)

# Optional: Machine Learning specific
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# Optional: GPU support
# torch>=2.0.0+cu118
# torchvision>=0.15.0+cu118

# Optional: Advanced embeddings
# InstructorEmbedding>=1.0.0

# Optional: Vector databases (alternatives)
# chromadb>=0.4.0
# qdrant-client>=1.0.0
# weaviate-client>=4.0.0

# Optional: Monitoring and observability
# opentelemetry-api>=1.20.0
# opentelemetry-sdk>=1.20.0

# Optional: Caching and performance
# redis>=4.5.0
# memcached>=1.6.0

# Optional: Authentication and security
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.0
# python-multipart>=0.0.6
